Export skirt/hair physics as PMX rigid bodies + joints from CySpring
CySpringPhysicsExporter converts the game's CySpring spring bones into MMD physics:
kinematic colliders on body/legs, dynamic sphere bodies on swaying bones, joints along
each chain with rotation limits, so babylon-mmd simulates skirt/hair/tail at runtime
instead of them clipping. Collider radii are clamped (CySpring runtime radii come back
inconsistently scaled). Also fixes PMXWriter writing rigid-body CollisionGroup as 4 bytes
instead of 1 (never hit before since physics was always empty).

+++ Assets/Scripts/Exporters/UnityPMXRuntimeLoader/LibMMD/Writer/PMXWriter.cs
@@ -93,7 +93,7 @@ namespace LibMMD.Writer
                 MMDReaderWriteUtil.WriteSizedString(writer, rigidBody.Name, pmxConfig.Encoding); // Name
                 MMDReaderWriteUtil.WriteSizedString(writer, rigidBody.NameEn, pmxConfig.Encoding); // NameEn
                 MMDReaderWriteUtil.WriteIndex(writer, rigidBody.AssociatedBoneIndex, pmxConfig.BoneIndexSize); // AssociatedBoneIndex
-                writer.Write(rigidBody.CollisionGroup); // CollisionGroup
+                writer.Write((byte)rigidBody.CollisionGroup); // CollisionGroup (1 byte; reader uses ReadByte)
                 writer.Write(rigidBody.CollisionMask); // CollisionMask
                 writer.Write((byte)rigidBody.Shape); // Shape
                 MMDReaderWriteUtil.WriteRawCoordinateVector3(writer, rigidBody.Dimemsions); // Dimemsions
